+++ packages/docs/src/pages/QuickStartPage.vue
@@ -1,6 +1,7 @@
 <script setup lang="ts">
 import { ZTitle, ZParagraph, ZCode } from '@kenconnet666/zui-vue'
 import ApiTable from '../components/ApiTable.vue'
+import CodeBlock from '../components/CodeBlock.vue'
 </script>
 
 <template>
@@ -10,15 +11,14 @@ import ApiTable from '../components/ApiTable.vue'
 
     <ZTitle :level="2">安装</ZTitle>
     <ZParagraph>使用 pnpm（推荐）、npm 或 yarn 安装：</ZParagraph>
-    <ZCode :inline="false" lang="bash" code="pnpm add @kenconnet666/zui-vue" />
+    <CodeBlock lang="bash" code="pnpm add @kenconnet666/zui-vue" />
 
     <ZTitle :level="2">注册 Provider</ZTitle>
     <ZParagraph>
       在应用根组件用 <ZCode code="ZBox" /> 包裹，注入主题。
       <ZCode code="ZBox" /> 是唯一必须的 Provider，它通过 CSS 变量和 Emotion 全局 class 向下传递主题上下文。
     </ZParagraph>
-    <ZCode
-      :inline="false"
+    <CodeBlock
       lang="vue"
       :code="`<script setup lang=&quot;ts&quot;>
 import { ZBox, zuiLight } from '@kenconnet666/zui-vue'
@@ -32,8 +32,7 @@ import { ZBox, zuiLight } from '@kenconnet666/zui-vue'
     />
 
     <ZTitle :level="2">在组件中使用</ZTitle>
-    <ZCode
-      :inline="false"
+    <CodeBlock
       lang="vue"
       :code="`<script setup lang=&quot;ts&quot;>
 import { ZButton, ZTitle } from '@kenconnet666/zui-vue'
@@ -50,8 +49,7 @@ import { ZButton, ZTitle } from '@kenconnet666/zui-vue'
       切换 <ZCode code=":theme" /> 为 <ZCode code="zuiDark" /> 即可全站暗色。 也可以动态绑定 ref
       实现亮暗切换。
     </ZParagraph>
-    <ZCode
-      :inline="false"
+    <CodeBlock
       lang="vue"
       :code="`<script setup lang=&quot;ts&quot;>
 import { computed, ref } from 'vue'
@@ -73,8 +71,7 @@ const theme = computed(() => dark.value ? zuiDark : zuiLight)
       通过 <ZCode code="theme.extend()" /> + TypeScript 声明合并，零模板注入自己的 brand token，
       全量 IDE 补全自动生效。
     </ZParagraph>
-    <ZCode
-      :inline="false"
+    <CodeBlock
       lang="ts"
       :code="`// zui.d.ts
 declare module '@kenconnet666/zui-vue' {